cornerRadius changed to float, removed label - related things from TextView, added data binding and row components to samples

+++ carbon/src/main/java/carbon/shadow/ShadowGenerator.java
@@ -94,7 +94,7 @@ public class ShadowGenerator {
         CornerView cornerView = (CornerView) view;
 
         int e = (int) Math.ceil(elevation);
-        int c = Math.max(e, cornerView.getCornerRadius());
+        int c = (int) Math.max(e, cornerView.getCornerRadius());
 
         for (Object o : shadowSet) {
             Shadow s = (Shadow) o;
